Reduce redunancies in code and fix bug in result screen
Prior this push there was a lot of repetitions inside the if-else statements of the functions. The result window now ask player to play again. In the html file were added the initial images. Next step improving CSS by adding a window that shows the player choice and the computer choice.